requirements.txt
sqlalchemy
logparser
scrapyd
scrapyd-client
psycopg2-binary
pyyaml